Create FibonacciNumber.java
This is a Java coded version of the answer.
This counts the value of the user given Fibonacci Number. It is done both with a loop and with a recursive method. It does not have any temporary integer variable when updating 'Previous' number.
Example: User gives value 10
Answer: The number corresponding to the 10th number in the Fibonacci sequence is: 55